// public/app/plugins/panel/textng/panelcfg.gen.ts
// Code generated - EDITING IS FUTILE. DO NOT EDIT.
//
// Generated by:
//     public/app/plugins/gen.go
// Using jennies:
//     TSTypesJenny
//     PluginTsTypesJenny
//
// Run 'make gen-cue' from repository root to regenerate.

// Generated from public/app/plugins/panel/textng/panelcfg.cue file.

export enum TextMode {
  Code = 'code',
  HTML = 'html',
  Markdown = 'markdown',
}

export enum CodeLanguage {
  Json = 'json',
  Plaintext = 'plaintext',
}

export const defaultCodeLanguage: CodeLanguage = CodeLanguage.Plaintext;

export interface CodeOptions {
  /**
   * The language passed to the CodeMirror editor
   */
  language: CodeLanguage;
  showLineNumbers: boolean;
}

export const defaultCodeOptions: Partial<CodeOptions> = {
  language: CodeLanguage.Plaintext,
  showLineNumbers: false,
};

export interface Options {
  code?: CodeOptions;
  content: string;
  mode: TextMode;
  /**
   * Soft-wrap long lines in the editor
   */
  wordWrap: boolean;
}

export const defaultOptions: Partial<Options> = {
  content: `# Title

For markdown syntax help: [commonmark.org/help](https://commonmark.org/help/)`,
  mode: TextMode.Markdown,
  wordWrap: true,
};